+++ src/components/Footer/FooterText.jsx
