--- src/components/Roster.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function TestMe() {
-  return <h1>Test</h1>;
-}
