+++ src/pages/Home.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 
 // importing website components
-import Football from "../components/Football";
-import Othersport from "../components/Othersports";
-import Basketball from "../components/Basketball";
+import Football from "../components/Football/Football";
+import Othersport from "../components/Othersports/Othersports";
+import Basketball from "../components/Basketball/Basketball";
 
 export default function Home() {
   return (
